+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Future
 #### NEW
 - added `config.html.tocLimit` to limit the levels of headers used in the toc
+- fixed html entities not showing as they should #14
 
 #### FIX
 - ...